==> code/04_creating_networks_calculating_network_metrics.Rmd <==
---
title: "Creating networks for each experimental period (pre and post) and calculating network metrics"
output: html_document
knit: (function(inputFile, encoding) {
      out_dir <- "knitted_markdown_files";
      rmarkdown::render(inputFile,
                        encoding=encoding,
                        output_dir=file.path(dirname(inputFile), out_dir))})
---

**Overall goal**

Make a network for each...

- Experimental replicate (year + patch combination),
- Experimental period (pre and post), and
- Sampling method (camera and pollen).

Visualize the network and calculate network metrics of interest. These network metrics, chosen *a priori*, are:

- Specialization metrics at the network level (two-dimensional Shannon diversity: H2')
- Specialization metrics at the species level (standardized Kullback-Leibler distance: d', species specificity index: SSI)
- Turnover metrics (Whittaker's beta diversity)
- Sampling completeness (Chao1, ACE)

```{r setup, include = FALSE}
library(dplyr)
library(tidyr)
library(purrr)
library(bipartite)

#knitr::opts_chunk$set(tidy.opts=list(width.cutoff=80), tidy=TRUE) #If want to knit to PDF, need to worry about text getting cut off
```

## Step 1: Import individual datsets (camera and pollen) and combine

Data have already been summarized to level of pre/post in previous scripts (01-02)
```{r step1, message = FALSE}
#Camera (visitation) data
#1. Remove unneeded columns. Number of flowers is mean number of flowers on camera, but it's not currently used in this analysis. Also, I will use the scientific names when visualizing networks, not the species codes, so will go ahead and replace those
#2. Standardize column names between the two datasets
#3. When plotting networks, it will be difficult to italicize the "sp." for plant species with unknown species, so just remove it now
#4. Convert interaction frequency to integer and round. This is for calculating d', which can only be calculated from integers
#5. Reorder columns
data_cameras <- read.csv("../data/export/for_analysis/Camera_data_summarized_for_pp_networks.csv") %>%
  select(-c(subset, flowers, plant_species, bird_species)) %>%
  rename(interaction_frequency = sightings_per_hour, plant_species = plant_scientific_name, bird_species = bird_scientific_name) %>%
  mutate(plant_species = ifelse(plant_species == "Drymonia sp.", "Drymonia", plant_species)) %>%
  mutate(plant_species = ifelse(plant_species == "Renealmia sp.", "Renealmia", plant_species)) %>%
  mutate(interaction_frequency = round(interaction_frequency * 10000, 0)) %>% 
  select(year:exp_phase, bird_species, plant_species, interaction_frequency, sightings, hours, everything()) %>%
  mutate(sampling_method = "visitation")

#Pollen data
#1. Remove unneeded columns
#2. Standardize column names between the two datasets
#3. Standardize names of experimental periods (capture_1/capture_2 -> pre/post)
data_pollen <- read.csv("../data/export/for_analysis/Pollen_data_summarized_for_pp_networks.csv") %>%
  select(-c(bird_species)) %>%
  rename(interaction_frequency = num_slides, plant_species = pollen_morphotype, bird_species = bird_scientific_name) %>%
  filter(exp_phase == "capture_1" | exp_phase == "capture_2") %>%
  mutate(exp_phase = ifelse(exp_phase == "capture_1", "pre", "post")) %>%
  mutate(sampling_method = "pollen")

#Make sure column names match up
head(data_cameras)
head(data_pollen)

#Combine data
network_data <- data_cameras %>%
  bind_rows(data_pollen) %>%
  mutate(exp_phase = factor(exp_phase, levels = c("pre", "post"))) %>%
  mutate(year_patch = paste(year, patch, sep = "_")) %>%
  select(sampling_method, year, patch, year_patch, everything()) %>%
  arrange(sampling_method, year, patch, exp_phase)
```

## Step 2: Build networks and visualize networks

**Since I need to make so many different networks, I've written some functions that I'll apply using purrr:map()**
```{r step2a, message = FALSE}
source("../code/helper_functions/Build_and_visualize_networks_for_experiment.R")
source("../code/helper_functions/Extract_data_from_safely_and_quietly_lists.R")

#To preserve side effects (error/warning messages) while running these functions, I'll wrap functions in purrr::safely() and quietly, then extract the results

#safely() returns results and errors
safe_matrix = safely(.f = create_matrix)
safe_specieslevel = safely(.f = calculate_specieslevel_metrics)

#quietly() returns results and warnings
quiet_networklevel = quietly(.f = calculate_networklevel_metrics) 
quiet_specieslevel = quietly(.f = calculate_specieslevel_metrics)
```

```{r step2b, include = FALSE}
#Test/demo the network functions
test_data_camera = subset_data(network_data, method_name = "visitation", year_name = "2016", patch_name = "24", exp_phase_name = "pre")
test_data_pollen = subset_data(network_data, method_name = "pollen", year_name = "2016", patch_name = "24", exp_phase_name = "pre")

(test_matrix_camera = create_matrix(test_data_camera, interaction_column = "interaction_frequency"))
(test_matrix_pollen = create_matrix(test_data_pollen, interaction_column = "interaction_frequency"))

calculate_networklevel_metrics(test_matrix_camera, index_list = c("H2"))
calculate_networklevel_metrics(test_matrix_pollen, index_list = c("H2"))

calculate_specieslevel_metrics(test_matrix_camera, index_list = c("d", "species specificity"))
calculate_specieslevel_metrics(test_matrix_pollen, index_list = c("d", "species specificity"))

visualize_network(test_matrix_camera, method_name = "visitation", label_heto = TRUE)
visualize_network(test_matrix_pollen, method_name = "pollen", label_heto = TRUE)
```

**Build and visualize networks**
```{r step2c, message = FALSE, warning = FALSE, results = 'hide', fig.show = 'hide'}
#This table will be filled in with networks, one network per row
base_table <- network_data %>%
  distinct(sampling_method, year, patch, year_patch, control_treatment, exp_phase) %>% 
  tibble()

#1. Subset data to replicate, period, and sampling method
#2. Create matrix
#3. Remove networks any without matrices. For some reason, is.null doesn't work (gets rid of all rows)
#4. Visualize networks,  using recordPlot to save the graphic
#5. Remove unneeded columns
network_results <- base_table %>%
  mutate(data = pmap(list(sampling_method, year, patch, exp_phase), ~subset_data(data = network_data, method_name = ..1, year_name = ..2, patch_name = ..3, exp_phase_name = ..4))) %>%
  mutate(matrix_safely = map(data, ~safe_matrix(data = ., interaction_column = "interaction_frequency")),
         matrix_error = map(matrix_safely, get_error),
         matrix = map(matrix_safely, get_result)) %>%
  filter(lengths(matrix_error) == 0) %>%
  mutate(network_plot = map2(matrix, sampling_method, ~visualize_network(matrix = .x, method_name = .y, label_heto = TRUE))) %>%
  mutate(row_id = row_number()) %>%
  mutate(plotname = paste(paste(row_id, sampling_method, year, patch, exp_phase, sep = "_"), ".png", sep = "")) %>%
  select(-contains("safely"), -contains("quietly"))
```

## Step 3: Calculate specialization metrics

**Calculate specialization metrics (network-level and species-level)**
```{r step3a, message = FALSE}
#1. Calculate network-level metrics and extract warnings
#2. Calculate species-level metrics and extract warnings + errors
#3. Remove unneeded columns
specialization_results <- network_results %>%
  mutate(networklevel_quietly = map(matrix, ~quiet_networklevel(matrix = .x, index_list = c("H2"))),
         networklevel_warning = map(networklevel_quietly, get_warning),
         networklevel = map(networklevel_quietly, get_result)) %>%
  mutate(specieslevel_safely = map(matrix, ~quiet_specieslevel(matrix = .x, index_list = c("d", "species specificity"))),
         specieslevel_quietly = map(matrix, ~safe_specieslevel(matrix = .x, index_list = c("d", "species specificity"))),
         specieslevel_error = map(specieslevel_safely, get_error),
         specieslevel_warning = map(specieslevel_quietly, get_warning),
         specieslevel = map(specieslevel_safely, get_result)) %>%
  select(-contains("safely"), -contains("quietly"))

#Review any errors/warnings
networklevel_warnings <- specialization_results %>%
  filter(grepl("Web is really too small", networklevel_warning, fixed = FALSE))

specieslevel_errors <- specialization_results %>%
  filter(lengths(specieslevel_error) != 0)

specieslevel_warnings <- specialization_results %>%
  filter(lengths(specieslevel_warning) != 0)
```

**Extract and organize specialization results**

Extract results and split into (1) network-level and (2) species-level
```{r step3b, message = FALSE}
#1. Remove results for networks that only have one plant or bird species
#2. Unnest results
#3. Convert to long form (because will join to *mean* specieslevel results in next step)
networklevel_results <- specialization_results %>%
  filter(!grepl("Web is really too small", networklevel_warning, fixed = FALSE)) %>%
  select(sampling_method, year, patch, control_treatment, exp_phase, networklevel) %>%
  unnest(networklevel) %>%
  rename(H2 = networklevel) %>%
  pivot_longer(cols = c(H2), names_to = "metric")

#At level of individual hummingbird species
#1. Remove results for networks that only have one plant or bird species
#2. Unnest results
specieslevel_results <- specialization_results %>%
  filter(!grepl("Web is really too small", networklevel_warning, fixed = FALSE)) %>%
  mutate(specieslevel = map(specieslevel, ~tibble::rownames_to_column(., var = "bird_species"))) %>%
  select(sampling_method, year, patch, control_treatment, exp_phase, specieslevel) %>%
  unnest(specieslevel)
```

Average species-level metrics across all hummingbird species and hummingbirds that are Heliconia specialists (i.e., green hermits and violet sabrewings)
```{r step3c, message = FALSE}
#Average values of d' and proportional generality for ALL HUMMINGBIRD SPECIES and GREH/VISA
mean_specieslevel_results_all <- specieslevel_results %>%
  group_by(sampling_method, year, patch, control_treatment, exp_phase) %>%
  summarise(across(c(d, species.specificity.index), ~mean(., na.rm = TRUE))) %>%
  ungroup() %>%
  mutate(bird_group = "all_spp")

mean_specieslevel_results_gv <- specieslevel_results %>%
  filter(bird_species == "Phaethornis guy" | bird_species == "Campylopterus hemileucurus") %>%
  group_by(sampling_method, year, patch, control_treatment, exp_phase) %>%
  summarise(across(c(d, species.specificity.index), ~mean(., na.rm = TRUE))) %>%
  ungroup() %>%
  mutate(bird_group = "greh_visa")

#Combine and convert to long-form (different columns for d, SSI -> different rows)
mean_specieslevel_results <- mean_specieslevel_results_all %>%
  bind_rows(mean_specieslevel_results_gv) %>%
  pivot_longer(cols = c(d, species.specificity.index), names_to = "metric")
```

Combine H2, mean d', and mean SSI results. Flag replicates that do not have network metrics pre and post.
```{r step3d, message = FALSE}
#Function to ID networks missing their partner
identify_unpaired_networks <- function(data){
  
  #Number of phases = number of experimental periods.
  #If have data for pre and post, should be 2 (not 1)
  unpaired_data <- data %>%
    summarise(num_phases = n()) %>%
    ungroup() %>%
    filter(num_phases == 1) %>%
    rename(unpaired = num_phases)
    
  return(unpaired_data)
  
}

#Combine data so far
specialization_results_for_analysis <- networklevel_results %>%
  mutate(bird_group = "all_spp") %>%
  bind_rows(mean_specieslevel_results)
  
#Identify unpaired networks
unpaired <- specialization_results_for_analysis %>%
  filter(!is.na(value)) %>%
  group_by(sampling_method, bird_group, metric, year, patch) %>%
  identify_unpaired_networks()
  
#Flag unpaired networks
specialization_results_for_analysis <- specialization_results_for_analysis %>%
  left_join(unpaired)
```

## Step 4: Calculate network-level interaction turnover (aka dissimilarity, aka pre-to-post beta diversity) and its additive components

Overall interaction turnover (WN = 'whole network') can be partitioned into two different subcomponents:

- OS = dissimilarity from rewiring among shared species only (OS = 'only shared')
- ST = dissimilarity from community composition (ST = 'species turnover')

Here, I implement the partitioning using bipartite::betalinkr(), which is a newer version of the method proposed by Poisot et al. 2012. 

The betalinkr function is developed by Jochen Fruend and partitions dissimilarity into additive components using the argument partitioning = "commondenom". If you want to ONLY look at rewiring between shared species, use partitioning = "poisot" (and ignore species turnover component).

**Prepare to calculate turnover metrics, binary and quantitative**
```{r step4a, message = FALSE}
#1. Make a table where pre matrix and post matrix each have a column
#2. Remove replicates that do not have networks pre AND post
#3. Prepare to compute binary AND quantitative indices
binary = c(TRUE, FALSE)

base_table_turnover <- network_results %>%
  select(sampling_method, year, patch, year_patch, control_treatment, exp_phase, matrix) %>% 
  pivot_wider(values_from = matrix, names_from = exp_phase) %>%
  mutate(include_replicate = map2(pre, post, ~ifelse(is.null(.x) | is.null(.y), "no", "yes"))) %>%
  filter(include_replicate == "yes") %>%
  select(-include_replicate) %>%
  crossing(binary) %>%
  select(binary, everything()) %>%
  arrange(binary)
```

**Calculate pre-to-post interaction turnover and its additive components**

Notes about betalinkr:

- If partitioning="commondenom" (and thus no external function is used), argument has to be sorensen (for Sorensen or Bray-Curtis dissimilarity index) or jaccard (for Jaccard or Ruzicka); bray is also allowed and gives the same result as sorensen.

- If only interested in OS (and want to ignore ST), can use partitioning = "poisot" (I'll go ahead and calculate both just in case)

- binary=TRUE is default. Seems like betalinkr function automatically converts to proportions when binary = FALSE

```{r step4b, message = FALSE, warning = FALSE}
turnover_results <- base_table_turnover %>%
  mutate(array = map2(pre, post, ~webs2array(list(.x, .y)))) %>%
  mutate(betalink_fruend = map2(array, binary, ~betalinkr(.x, partitioning = "commondenom", index = "bray", partition.st = FALSE, partition.rr = FALSE, binary = .y))) %>%
  mutate(betalink_poisot = map2(array, binary, ~betalinkr(.x, partitioning = "poisot", index = "bray", partition.st = FALSE, partition.rr = FALSE, binary = .y, distofempty = "na")))
```

**Extract results**

```{r step4c, message = FALSE}
#Fruend method, partitioned into different components
turnover_results_fruend <- turnover_results %>%
  mutate(betalink_fruend = map(betalink_fruend, ~as.data.frame(.))) %>%
  mutate(betalink_fruend = map(betalink_fruend, ~tibble::rownames_to_column(., var = "component"))) %>%
  select(binary, sampling_method, year, patch, control_treatment, betalink_fruend) %>%
  unnest(betalink_fruend) %>%
  rename(value = ".") %>%
  pivot_wider(names_from = "component", values_from = "value") %>%
  select(-S)

turnover_results_for_analysis <- turnover_results_fruend
```

```{r eval = FALSE, include = FALSE}
#Did not end up using this method because we were interested in additive partitions
#Poisot method
turnover_results_poisot <- turnover_results %>%
  mutate(betalink_poisot = map(betalink_poisot, ~as.data.frame(.))) %>%
  mutate(betalink_poisot = map(betalink_poisot, ~tibble::rownames_to_column(., var = "component"))) %>%
  select(binary, sampling_method, year, patch, control_treatment, betalink_poisot) %>%
  unnest(betalink_poisot) %>%
  rename(value = ".") %>%
  pivot_wider(names_from = "component", values_from = "value") %>%
  select(-WN, -ST, -S) %>% #Ignore these columns
  rename(OS.poisot = OS)
  
#Combine results from different methods
turnover_results_for_analysis <- turnover_results_fruend %>%
  left_join(turnover_results_poisot)
```

## Step 5: Calculate network-level sampling completeness

To calculate sampling completeness, I need actual numbers of observations - NOT rates, and NOT the rate converted to an integer (x 10,000). So for the camera data, the interaction frequency needs to be the number of sightings. For the pollen data, number of slides is fine; this the interaction frequency used in any calculations so far.

```{r step5, message = FALSE}
#Re-make network results for camera data using number of sightings, not sighting rate
network_results_camera <- base_table %>%
  filter(sampling_method == "visitation") %>%
  mutate(data = pmap(list(sampling_method, year, patch, exp_phase), ~subset_data(data = network_data, method_name = ..1, year_name = ..2, patch_name = ..3, exp_phase_name = ..4))) %>%
  mutate(matrix_safely = map(data, ~safe_matrix(data = ., interaction_column = "sightings")),
         matrix_error = map(matrix_safely, get_error),
         matrix = map(matrix_safely, get_result)) %>%
  filter(lengths(matrix_error) == 0)

#1. Replace camera networks in original network results
#2. Calculate sampling completeness (automatically returns two metrics: Chao1 and ACE)
completeness_results <- network_results %>%
  filter(sampling_method != "visitation") %>%
  select(-c(network_plot, contains("name"))) %>%
  bind_rows(network_results_camera) %>%
  mutate(completeness = map(matrix, calculate_completeness))

#Extract results
completeness_results_for_analysis <- completeness_results %>%
  mutate(completeness = map(completeness, ~as.data.frame(.))) %>%
  mutate(completeness = map(completeness, ~tibble::rownames_to_column(., var = "component"))) %>%
  select(sampling_method, year, patch, year_patch, control_treatment, exp_phase, completeness) %>%
  unnest(completeness) %>%
  rename(value = "V1") %>%
  pivot_wider(names_from = "component", values_from = "value") %>%
  mutate(completeness.chao1 = S.obs/S.chao1) %>%
  mutate(completeness.ACE = S.obs/S.ACE)
```

## Step 6: Export results

```{r step6}
write.csv(specialization_results_for_analysis, "../data/export/for_analysis/Network_specialization_for_analysis.csv", row.names = FALSE)
write.csv(turnover_results_for_analysis, "../data/export/for_analysis/Network_dissimilarity_for_analysis.csv", row.names = FALSE)
write.csv(completeness_results_for_analysis, "../data/export/for_analysis/Sampling_completeness_for_analysis.csv", row.names = FALSE)

path_name <- c("../results/network_visualizations/pre_post_networks_experiment")
walk2(network_results$plotname, network_results$network_plot, ~ggplot2::ggsave(filename = .x, path = path_name, plot = .y, height = 12, width = 10, bg = "white", device = "png"))
```
